backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.7.1
pymongo>=4.9,<5
pydantic[email]>=2.11.7,<3.0.0
python-dotenv==1.0.1
certifi==2025.1.31
